手动创建采购订单

+++ common-base/src/main/java/com/upedge/common/utils/IdGenerate.java
@@ -44,7 +44,7 @@ public class IdGenerate {
 		return Math.abs(random.nextLong());
 	}
 
-	public static long randomInt() {
+	public static int randomInt() {
 
 		return Math.abs(random.nextInt());
 	}
